requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0


langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.51.0


qdrant-client>=1.11.0


pandas>=2.2.0
openpyxl>=3.1.0

ipykernel>=6.29.0


python-dotenv>=1.0.0

langgraph-cli[inmem]